(1) Grid mask is applied in the source filter of clients: unmasked values are replaced by NaN. It is not reconstructed any more by servers.

(2) Domain/axis mask has been incorporated into data index, with only data index sent to servers.


git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS/trunk@1637 d613d254-d18a-4304-b296-5125d4c4d670

+++ src/node/axis.hpp
@@ -125,7 +125,7 @@ namespace xios {
         bool hasBounds;
         bool hasLabel;
 
-        CArray<size_t,1> localIndexToWriteOnServer;        
+        CArray<int,1> localIndexToWriteOnServer;
 
       private:
          void checkData();
